[138] Move all stores into /stores directory
