+++ src/stores/OrderFilledCancelledLazyStore.ts
@@ -1,6 +1,6 @@
 import BigNumber from 'bignumber.js';
 import * as _ from 'lodash';
-import { ContractWrapper } from './contract_wrappers/ContractWrapper';
+import { ContractWrapper } from '../contract_wrappers/ContractWrapper';
 
 interface InstanceFilledCancelledStore {
   filledOrCancelledQty: {
